refactor: (#61) 타입, 상수 관련 팀 코드 컨벤션 변경으로 인한 수정
