+++ frontend/__test__/hooks/useSelect.test.tsx
@@ -3,14 +3,14 @@ import { act } from 'react-dom/test-utils';
 
 import { useSelect } from '@hooks/useSelect';
 
-import { PostStatusType } from '@components/post/PostListPage/constants/option';
+import { PostStatus } from '@components/post/PostListPage/types';
 
 const INIT_SELECTED_OPTION = 'progress';
 const CHANGE_SELECTED_OPTION = 'closed';
 
 describe('useSelect 훅이 의도한대로 동작하는지 테스트한다.', () => {
   test('초기 값이 의도한대로 설정 되었는지 확인한다.', () => {
-    const { result } = renderHook(() => useSelect<PostStatusType>(INIT_SELECTED_OPTION));
+    const { result } = renderHook(() => useSelect<PostStatus>(INIT_SELECTED_OPTION));
 
     const { selectedOption } = result.current;
 
@@ -20,7 +20,7 @@ describe('useSelect 훅이 의도한대로 동작하는지 테스트한다.', ()
   test('값을 바꾸는 함수가 값을 바꿨는지 확인한다.', () => {
     const INIT_SELECTED_OPTION = 'progress';
 
-    const { result } = renderHook(() => useSelect<PostStatusType>(INIT_SELECTED_OPTION));
+    const { result } = renderHook(() => useSelect<PostStatus>(INIT_SELECTED_OPTION));
 
     const { handleOptionChange } = result.current;
